refactor: rename calltree.nvim to litee.nvim
this commit changes the name of Calltree.nvim to
LITEE.nvim and updates all the necessary code points
to reflect this change.
